+++ server/component/style/carousel/tpl_caption.php
@@ -1,0 +1,3 @@
+<div class="carousel-caption d-none d-md-block">
+    <p><?php echo $caption; ?></p>
+</div>