+++ server/component/style/carousel/tpl_carousel.php
@@ -1,4 +1,4 @@
-<div id="<?php echo $this->id_prefix; ?>-carousel" class="carousel slide" data-ride="carousel">
+<div id="<?php echo $this->id_prefix; ?>-carousel" class="carousel slide <?php echo $crossfade; ?> <?php echo $this->css; ?>" data-ride="carousel">
     <?php $this->output_indicator_wrapper(); ?>
     <div class="carousel-inner">
         <?php $this->output_carousel_items(); ?>